update example to minimal
