+++ with-formik/App.js
@@ -1,21 +1,124 @@
-import * as React from 'react';
-import { NavigationContainer } from '@react-navigation/native';
-import { createStackNavigator } from '@react-navigation/stack';
+import React from 'react';
+import {
+  View,
+  Text,
+  TouchableOpacity,
+  TextInput,
+  StyleSheet,
+  Dimensions
+} from 'react-native';
+import { Formik } from 'formik';
+import * as Yup from 'yup';
 
-import LoginScreen from './src/screens/LoginScreen';
-import SignupScreen from './src/screens/SignupScreen';
-import HomeScreen from './src/screens/HomeScreen';
+const validationSchema = Yup.object().shape({
+  email: Yup.string()
+    .label('Email')
+    .email('Enter a valid email')
+    .required('Please enter a registered email'),
+  password: Yup.string()
+    .label('Password')
+    .required()
+    .min(6, 'Password must have at least 6 characters ')
+});
 
-const Stack = createStackNavigator();
+const ErrorMessage = ({ errorValue }) => (
+  <View style={styles.errorContainer}>
+    <Text style={styles.errorText}>{errorValue}</Text>
+  </View>
+);
 
 export default function App() {
+  function onLoginHandler(values) {
+    const { email, password } = values;
+
+    alert(`Credentials entered. email: ${email}, password: ${password}`);
+  }
+
   return (
-    <NavigationContainer>
-      <Stack.Navigator>
-        <Stack.Screen name="Login" component={LoginScreen} />
-        <Stack.Screen name="Signup" component={SignupScreen} />
-        <Stack.Screen name="Home" component={HomeScreen} />
-      </Stack.Navigator>
-    </NavigationContainer>
+    <View style={styles.container}>
+      <Formik
+        initialValues={{ email: '', password: '' }}
+        onSubmit={(values, actions) => {
+          onLoginHandler(values, actions);
+        }}
+        validationSchema={validationSchema}
+      >
+        {({
+          handleChange,
+          values,
+          errors,
+          touched,
+          handleSubmit,
+          handleBlur
+        }) => (
+          <>
+            <TextInput
+              style={styles.input}
+              numberOfLines={1}
+              value={values.email}
+              placeholder="Enter email"
+              onChangeText={handleChange('email')}
+              autoCapitalize="none"
+              onBlur={handleBlur('email')}
+            />
+            <ErrorMessage errorValue={touched.email && errors.email} />
+            <TextInput
+              style={styles.input}
+              numberOfLines={1}
+              value={values.password}
+              placeholder="Enter password"
+              onChangeText={handleChange('password')}
+              autoCapitalize="none"
+              onBlur={handleBlur('password')}
+              secureTextEntry={true}
+            />
+            <ErrorMessage errorValue={touched.password && errors.password} />
+            <TouchableOpacity
+              onPress={handleSubmit}
+              style={styles.buttonContainer}
+            >
+              <Text style={styles.buttonText}>Login</Text>
+            </TouchableOpacity>
+          </>
+        )}
+      </Formik>
+    </View>
   );
 }
+
+const styles = StyleSheet.create({
+  errorContainer: {
+    marginVertical: 5
+  },
+  errorText: {
+    color: 'red'
+  },
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    marginTop: 40
+  },
+  input: {
+    marginVertical: 10,
+    width: Dimensions.get('window').width - 100,
+
+    height: 40,
+    borderWidth: 1,
+    padding: 10,
+    borderRadius: 5
+  },
+  buttonContainer: {
+    marginVertical: 10,
+    justifyContent: 'center',
+    alignItems: 'center',
+    padding: 10,
+    width: Dimensions.get('window').width - 200,
+    height: 44,
+    borderRadius: 5,
+    backgroundColor: '#343434'
+  },
+  buttonText: {
+    fontSize: 18,
+    color: '#ffffff'
+  }
+});
