+++ with-formik/package.json
@@ -1,18 +1,10 @@
 {
   "dependencies": {
-    "@react-native-community/masked-view": "0.1.6",
-    "@react-navigation/native": "5.4.3",
-    "@react-navigation/stack": "5.4.0",
     "expo": "37.0.11",
-    "firebase": "7.9.0",
     "formik": "2.1.4",
     "react": "16.9.0",
     "react-dom": "16.9.0",
     "react-native": "https://github.com/expo/react-native/archive/sdk-37.0.1.tar.gz",
-    "react-native-gesture-handler": "1.6.1",
-    "react-native-reanimated": "~1.7.0",
-    "react-native-safe-area-context": "0.7.3",
-    "react-native-screens": "~2.2.0",
     "react-native-web": "0.11.7",
     "yup": "0.29.0"
   },
