--- with-formik/src/config/firebaseConfig.js
@@ -1,7 +1,0 @@
-export default {
-  apiKey: 'AIzaSyClsBKWOOXfzeb48Rm80hX09X87J_TB4mM',
-  authDomain: 'rnexpo0.firebaseapp.com',
-  databaseURL: 'https://rnexpo0.firebaseio.com',
-  storageBucket: '',
-  messagingSenderId: '256218572662'
-};
